Replaced cube_class with legacy cube everywere
