+++ temp/side4.txt
@@ -1,3 +1,3 @@
-2.000000000000000000e+00 2.000000000000000000e+00 2.000000000000000000e+00
-2.000000000000000000e+00 2.000000000000000000e+00 2.000000000000000000e+00
-2.000000000000000000e+00 2.000000000000000000e+00 2.000000000000000000e+00
+1.000000000000000000e+00 3.000000000000000000e+00 2.000000000000000000e+00
+0.000000000000000000e+00 3.000000000000000000e+00 3.000000000000000000e+00
+5.000000000000000000e+00 2.000000000000000000e+00 1.000000000000000000e+00
