+++ temp/side5.txt
@@ -1,3 +1,3 @@
-5.000000000000000000e+00 5.000000000000000000e+00 5.000000000000000000e+00
-5.000000000000000000e+00 5.000000000000000000e+00 5.000000000000000000e+00
-5.000000000000000000e+00 5.000000000000000000e+00 5.000000000000000000e+00
+5.000000000000000000e+00 5.000000000000000000e+00 4.000000000000000000e+00
+5.000000000000000000e+00 4.000000000000000000e+00 3.000000000000000000e+00
+2.000000000000000000e+00 5.000000000000000000e+00 3.000000000000000000e+00
